% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read}
\alias{fars_read}
\title{Import data from an external filename}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the location to import data.
The value of this argument should describe the full path of an external file.}
}
\value{
This function returns a data frame tbl.
}
\description{
This function imports data from an external CSV file.
You can customize which file to be read by using the argument \code{filename}.
}
\note{
This function will return an error when a file doesn't exist.
}
\examples{
fars_read("my_input.csv")

}
