% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Create a list for a vector of years.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{An integer or string vector that describes years from which a coressponding files will be read.}
}
\value{
This function returns a list of month and year for reading data.
}
\description{
This function reads each file from given years.
You can customize years to read by using the argument \code{years}.
}
\note{
This function will return an error if file doesn't exist with an error and warning message.
The error tells which file doesn't exist, while the warning indicates which year of the vector is invalid.
}
\examples{
\dontrun{
fars_read_years(c(2013:2015))
fars_read_years(c("2013","2015"))
}

}
