% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{make_filename}
\alias{make_filename}
\title{Create a file name from a given year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer or string scale giving a year from which the function will create a file name.}
}
\value{
This function returns a string.
}
\description{
This function create a file name from a given year.
You can customize a year to create by using the argument \code{year}.
}
\examples{
make_filename(2018)
make_filename("2018")

}
